+++ dotnet/Data/InMemoryMerchantSettingsRepository.cs
@@ -39,12 +39,12 @@
             this.SetMerchantSettings("FedEx", merchantSettings);
         }
 
-        public Task SetMerchantSettings(string carrier, MerchantSettings merchantSettings)
+        public Task<bool> SetMerchantSettings(string carrier, MerchantSettings merchantSettings)
         {
             this._inMemorySettings.Remove("settings");
             this._inMemorySettings.Add("settings", merchantSettings);
 
-            return Task.CompletedTask;
+            return Task.FromResult(true);
         }
 
         public Task<MerchantSettings> GetMerchantSettings(string carrier)